// backend-cuadrusistem/src/domain/ingrediente/ingrediente.enum.ts
export enum Ingrediente {
  PAN_MESA_SUPER_GIGAN = 'PAN_MESA_SUPER_GIGAN',
  VIENESA_PERSONAL = 'VIENESA_PERSONAL',

  // Bebestibles (se consumen directamente)
  COCA_COLA_LATA = 'COCA_COLA_LATA',
  COCA_COLA_591CC = 'COCA_COLA_591CC',
  AGUA_SIN_GAS = 'AGUA_SIN_GAS',

  // Empanadas / masas
  MASA_EMP_EMP = 'MASA_EMP_EMP',
  PINO_CARNE = 'PINO_CARNE',
  QUESO_EMP = 'QUESO_EMP',

  // Agrega aquí TODO ingrediente válido
}